+++ exonum/src/storage/proof_map_index/key.rs
@@ -22,7 +22,12 @@ pub const LEAF_KEY_PREFIX: u8 = 1;
 
 /// Size in bytes of the `ProofMapKey`.
 pub const KEY_SIZE: usize = HASH_SIZE;
-pub const DB_KEY_SIZE: usize = KEY_SIZE + 2;
+/// Size in bytes of the `ProofPath`.
+pub const PROOF_PATH_SIZE: usize = KEY_SIZE + 2;
+/// Position of the byte with kind of the `ProofPath`.
+pub const PROOF_PATH_KIND_POS: usize = 0;
+/// Position of the byte with total length of the branch.
+pub const PROOF_PATH_LEN_POS: usize = KEY_SIZE + 1;
 
 /// A trait that defines a subset of storage key types which are suitable for use with
 /// `ProofMapIndex`.
@@ -56,14 +61,6 @@ pub enum ChildKind {
     Right,
 }
 
-/// A struct that represents a bit slices of the proof map keys.
-#[derive(Clone)]
-pub struct DBKey {
-    data: [u8; KEY_SIZE],
-    from: u16,
-    to: u16,
-}
-
 impl ::std::ops::Not for ChildKind {
     type Output = ChildKind;
 
@@ -75,52 +72,94 @@ impl ::std::ops::Not for ChildKind {
     }
 }
 
-impl DBKey {
-    /// Create a new bit slice from the given binary data.
-    pub fn leaf<K: ProofMapKey>(key: &K) -> DBKey {
+/// A structure that represents paths to `ProofMapIndex` nodes.
+///
+/// # Binary representation
+///
+/// | Position in bytes     | Description                   	                    |
+/// |-------------------    |----------------------------------------------         |
+/// | 0               	    | `ProofPath` kind: (0 is branch, 1 is leaf)            |
+/// | 1..33                 | `ProofMapKey` bytes.    	                            |
+/// | 33                    | Total length in bits of `ProofMapKey` for branches.   |
+///
+#[derive(Copy, Clone)]
+pub struct ProofPath {
+    bytes: [u8; PROOF_PATH_SIZE],
+    start: u16,
+}
+
+impl ProofPath {
+    /// Create a path from the given key.
+    pub fn new<K: ProofMapKey>(key: &K) -> ProofPath {
         debug_assert_eq!(key.size(), KEY_SIZE);
 
-        let mut data = [0; KEY_SIZE];
-        key.write(&mut data);
-        DBKey {
-            data: data,
-            from: 0,
-            to: (KEY_SIZE * 8) as u16,
+        let mut data = [0; PROOF_PATH_SIZE];
+        data[0] = LEAF_KEY_PREFIX;
+        key.write(&mut data[1..KEY_SIZE + 1]);
+        data[PROOF_PATH_LEN_POS] = 0;
+        ProofPath::from_raw(data)
+    }
+
+    /// Checks if this is a path to a leaf `ProofMapIndex` node.
+    pub fn is_leaf(&self) -> bool {
+        self.bytes[0] == LEAF_KEY_PREFIX
+    }
+
+    /// Returns the byte representation of contained `ProofMapKey`.
+    pub fn as_bytes(&self) -> &[u8] {
+        &self.bytes
+    }
+
+    /// Constructs the `ProofPath` from raw bytes.
+    fn from_raw(raw: [u8; PROOF_PATH_SIZE]) -> ProofPath {
+        debug_assert!(
+            (raw[PROOF_PATH_KIND_POS] != LEAF_KEY_PREFIX) || (raw[PROOF_PATH_LEN_POS] == 0),
+            "ProofPath is inconsistent"
+        );
+
+        ProofPath {
+            bytes: raw,
+            start: 0,
         }
     }
 
-    #[doc(hidden)]
-    pub fn from(&self) -> u16 {
-        self.from
+    /// Sets the right border of the bit range.
+    fn set_end(&mut self, end: Option<u8>) {
+        // Update ProofPath kind and right bound.
+        if let Some(pos) = end {
+            self.bytes[0] = BRANCH_KEY_PREFIX;
+            self.bytes[PROOF_PATH_LEN_POS] = pos as u8;
+        } else {
+            self.bytes[0] = LEAF_KEY_PREFIX;
+            self.bytes[PROOF_PATH_LEN_POS] = 0;
+        };
+    }
+}
+
+/// The bits representation of the `ProofPath`.
+pub(crate) trait BitsRange {
+    /// Returns the left border of the range.
+    fn start(&self) -> u16;
+
+    /// Returns the right border of the range.
+    fn end(&self) -> u16;
+
+    /// Returns length in bits of the range.
+    fn len(&self) -> u16 {
+        self.end() - self.start()
     }
 
-    // TODO: terrible hack, try to remove this (ECR-22)
-    #[doc(hidden)]
-    pub fn set_from(&mut self, from: u16) {
-        self.from = from
+    /// Returns true if the range has zero length.
+    fn is_empty(&self) -> bool {
+        self.end() == self.start()
     }
 
-    #[doc(hidden)]
-    pub fn to(&self) -> u16 {
-        self.to
-    }
+    /// Get bit at index `idx`.
+    fn bit(&self, idx: u16) -> ChildKind {
+        debug_assert!(self.start() + idx < self.end());
 
-    /// Returns length of the `DBKey`.
-    pub fn len(&self) -> u16 {
-        self.to - self.from
-    }
-
-    /// Returns true if `DBKey` has zero length.
-    pub fn is_empty(&self) -> bool {
-        self.to == self.from
-    }
-
-    /// Get bit at position `idx`.
-    pub fn get(&self, idx: u16) -> ChildKind {
-        debug_assert!(self.from + idx < self.to);
-
-        let pos = self.from + idx;
-        let chunk = self.data[(pos / 8) as usize];
+        let pos = self.start() + idx;
+        let chunk = self.raw_key()[(pos / 8) as usize];
         let bit = pos % 8;
         let value = (1 << bit) & chunk;
         if value != 0 {
@@ -130,142 +169,105 @@ impl DBKey {
         }
     }
 
-    /// Shortens this DBKey to the specified length.
-    pub fn prefix(&self, suffix: u16) -> DBKey {
-        debug_assert!(self.from + suffix <= self.data.len() as u16 * 8);
+    /// Returns the copy of this bit range with the given left border.
+    fn start_from(&self, pos: u16) -> Self;
 
-        DBKey {
-            data: self.data,
-            from: self.from,
-            to: self.from + suffix,
-        }
+    /// Returns a copy of this bit range shortened to the specified length.
+    fn prefix(&self, len: u16) -> Self;
+
+    /// Returns the copy of this bit range where the start is shifted by the `len`
+    /// bits to the right.
+    fn suffix(&self, len: u16) -> Self;
+
+    /// Checks if this bit range contains the other bit range as a prefix,
+    /// provided that the start positions of both ranges are the same.
+    fn starts_with(&self, other: &Self) -> bool {
+        self.common_prefix_len(other) == other.len()
     }
 
-    /// Return object which represents a view on to this slice (further) offset by `i` bits.
-    pub fn suffix(&self, suffix: u16) -> DBKey {
-        debug_assert!(self.from + suffix <= self.to);
+    /// Returns the raw bytes of the key.
+    fn raw_key(&self) -> &[u8];
 
-        DBKey {
-            data: self.data,
-            from: self.from + suffix,
-            to: self.to,
-        }
-    }
-
-    /// Shortens this DBKey to the specified length.
-    pub fn truncate(&self, size: u16) -> DBKey {
-        debug_assert!(self.from + size <= self.to);
-
-        DBKey {
-            data: self.data,
-            from: self.from,
-            to: self.from + size,
-        }
-    }
-
-    /// Returns how many bits at the beginning matches with `other`
-    pub fn common_prefix(&self, other: &Self) -> u16 {
-        // We assume that all slices created from byte arrays with the same length
-        if self.from != other.from {
+    /// Returns the length of the common prefix between this and the other range,
+    /// provided that they start from the same position.
+    /// If start positions differ, returns 0.
+    fn common_prefix_len(&self, other: &Self) -> u16 {
+        if self.start() != other.start() {
             0
         } else {
-            let from = self.from / 8;
-            let to = min((self.to + 7) / 8, (other.to + 7) / 8);
+            let from = self.start() / 8;
+            let to = min((self.end() + 7) / 8, (other.end() + 7) / 8);
             let max_len = min(self.len(), other.len());
 
             for i in from..to {
-                let x = self.data[i as usize] ^ other.data[i as usize];
+                let x = self.raw_key()[i as usize] ^ other.raw_key()[i as usize];
                 if x != 0 {
                     let tail = x.trailing_zeros() as u16;
-                    return min(i * 8 + tail - self.from, max_len);
+                    return min(i * 8 + tail - self.start(), max_len);
                 }
             }
 
             max_len
         }
     }
-
-    /// Returns true if we starts with the same prefix at the whole of `Other`
-    pub fn starts_with(&self, other: &Self) -> bool {
-        self.common_prefix(other) == other.len()
-    }
-
-    /// Returns true if self.to not changed
-    pub fn is_leaf(&self) -> bool {
-        self.to == (KEY_SIZE * 8) as u16
-    }
-
-    // TODO: terrible hack, try to remove this (ECR-22)
-    /// Represents `DBKey` as byte array and returns it
-    pub fn as_bytes(&self) -> Box<[u8]> {
-        let mut buffer = Box::new([0u8; DB_KEY_SIZE as usize]);
-        self.write(buffer.as_mut());
-        buffer
-    }
 }
 
-impl AsRef<[u8]> for DBKey {
-    fn as_ref(&self) -> &[u8] {
-        &self.data
-    }
-}
-
-impl StorageKey for DBKey {
-    fn size(&self) -> usize {
-        DB_KEY_SIZE
+impl BitsRange for ProofPath {
+    fn start(&self) -> u16 {
+        self.start
     }
 
-    fn write(&self, buffer: &mut [u8]) {
+    fn end(&self) -> u16 {
         if self.is_leaf() {
-            buffer[0] = LEAF_KEY_PREFIX;
-            buffer[1..KEY_SIZE + 1].copy_from_slice(&self.data);
-            buffer[KEY_SIZE + 1] = 0;
+            KEY_SIZE as u16 * 8
         } else {
-            buffer[0] = BRANCH_KEY_PREFIX;
-            let right = (self.to as usize + 7) / 8;
-            buffer[1..right + 1].copy_from_slice(&self.data[0..right]);
-            if self.to % 8 != 0 {
-                buffer[right] &= !(255u8 << (self.to % 8));
-            }
-            for i in buffer.iter_mut().take(KEY_SIZE + 1).skip(right + 1) {
-                *i = 0
-            }
-            buffer[KEY_SIZE + 1] = self.to as u8;
+            u16::from(self.bytes[PROOF_PATH_LEN_POS])
         }
     }
 
-    fn read(buffer: &[u8]) -> Self {
-        debug_assert_eq!(buffer.len(), DB_KEY_SIZE);
-        let mut data = [0; KEY_SIZE];
-        data[..].copy_from_slice(&buffer[1..KEY_SIZE + 1]);
-        let to = match buffer[0] {
-            LEAF_KEY_PREFIX => KEY_SIZE as u16 * 8,
-            BRANCH_KEY_PREFIX => u16::from(buffer[DB_KEY_SIZE - 1]),
-            _ => unreachable!("wrong key prefix"),
-        };
-        DBKey {
-            data: data,
-            from: 0,
-            to: to,
-        }
+    fn start_from(&self, pos: u16) -> Self {
+        debug_assert!(pos <= self.end());
+
+        let mut key = ProofPath::from_raw(self.bytes);
+        key.start = pos;
+        key
+    }
+
+    fn prefix(&self, len: u16) -> Self {
+        let end = self.start + len;
+        let key_len = self.raw_key().len() as u16 * 8;
+        debug_assert!(end < key_len);
+
+        let mut key = ProofPath::from_raw(self.bytes);
+        key.start = self.start;
+        key.set_end(Some(end as u8));
+        key
+    }
+
+    fn suffix(&self, len: u16) -> Self {
+        self.start_from(self.start() + len)
+    }
+
+    fn raw_key(&self) -> &[u8] {
+        &self.bytes[1..KEY_SIZE + 1]
     }
 }
 
-impl PartialEq for DBKey {
+impl PartialEq for ProofPath {
     fn eq(&self, other: &Self) -> bool {
         self.len() == other.len() && self.starts_with(other)
     }
 }
 
-impl ::std::fmt::Debug for DBKey {
+impl ::std::fmt::Debug for ProofPath {
     fn fmt(&self, f: &mut ::std::fmt::Formatter) -> ::std::fmt::Result {
         // 8 bits + '|' symbol per byte.
         let mut bits = String::with_capacity(KEY_SIZE * 9);
-        for byte in 0..self.data.len() {
-            let chunk = self.data[byte];
+        for byte in 0..self.raw_key().len() {
+            let chunk = self.raw_key()[byte];
             for bit in (0..8).rev() {
                 let i = (byte * 8 + bit) as u16;
-                if i < self.from || i >= self.to {
+                if i < self.start() || i >= self.end() {
                     bits.push('_');
                 } else {
                     bits.push(if (1 << bit) & chunk == 0 { '0' } else { '1' });
@@ -274,20 +276,47 @@ impl ::std::fmt::Debug for DBKey {
             bits.push('|');
         }
 
-        f.debug_struct("DBKey")
-            .field("begin", &self.from)
-            .field("end", &self.to)
+        f.debug_struct("ProofPath")
+            .field("start", &self.start())
+            .field("end", &self.end())
             .field("bits", &bits)
             .finish()
     }
 }
 
+impl StorageKey for ProofPath {
+    fn size(&self) -> usize {
+        PROOF_PATH_SIZE
+    }
+
+    fn write(&self, buffer: &mut [u8]) {
+        buffer.copy_from_slice(&self.bytes);
+        // Cut of the bits that lie to the right of the end.
+        if !self.is_leaf() {
+            let right = (self.end() as usize + 7) / 8;
+            if self.end() % 8 != 0 {
+                buffer[right] &= !(255u8 << (self.end() % 8));
+            }
+            for i in buffer.iter_mut().take(KEY_SIZE + 1).skip(right + 1) {
+                *i = 0
+            }
+        }
+    }
+
+    fn read(buffer: &[u8]) -> Self {
+        debug_assert_eq!(buffer.len(), PROOF_PATH_SIZE);
+        let mut data = [0; PROOF_PATH_SIZE];
+        data.copy_from_slice(buffer);
+        ProofPath::from_raw(data)
+    }
+}
+
 #[test]
-fn test_dbkey_storage_key_leaf() {
-    let key = DBKey::leaf(&[250; 32]);
-    let mut buf = vec![0; DB_KEY_SIZE];
+fn test_proof_path_storage_key_leaf() {
+    let key = ProofPath::new(&[250; 32]);
+    let mut buf = vec![0; PROOF_PATH_SIZE];
     key.write(&mut buf);
-    let key2 = DBKey::read(&buf);
+    let key2 = ProofPath::read(&buf);
 
     assert_eq!(buf[0], LEAF_KEY_PREFIX);
     assert_eq!(buf[33], 0);
@@ -296,15 +325,15 @@ fn test_dbkey_storage_key_leaf() {
 }
 
 #[test]
-fn test_dbkey_storage_key_branch() {
-    let mut key = DBKey::leaf(&[255; 32]);
+fn test_proof_path_storage_key_branch() {
+    let mut key = ProofPath::new(&[255u8; 32]);
     key = key.prefix(11);
     key = key.suffix(5);
 
-    let mut buf = vec![0; DB_KEY_SIZE];
+    let mut buf = vec![0; PROOF_PATH_SIZE];
     key.write(&mut buf);
-    let mut key2 = DBKey::read(&buf);
-    key2.set_from(5);
+    let mut key2 = ProofPath::read(&buf);
+    key2.start = 5;
 
     assert_eq!(buf[0], BRANCH_KEY_PREFIX);
     assert_eq!(buf[33], 11);
@@ -314,141 +343,141 @@ fn test_dbkey_storage_key_branch() {
 }
 
 #[test]
-fn test_dbkey_suffix() {
-    let b = DBKey::read(b"\x00\x01\x02\xFF\x0C0000000000000000000000000000\x20");
+fn test_proof_path_suffix() {
+    let b = ProofPath::from_raw(*b"\x00\x01\x02\xFF\x0C0000000000000000000000000000\x20");
 
     assert_eq!(b.len(), 32);
-    assert_eq!(b.get(0), ChildKind::Right);
-    assert_eq!(b.get(7), ChildKind::Left);
-    assert_eq!(b.get(8), ChildKind::Left);
-    assert_eq!(b.get(9), ChildKind::Right);
-    assert_eq!(b.get(15), ChildKind::Left);
-    assert_eq!(b.get(16), ChildKind::Right);
-    assert_eq!(b.get(20), ChildKind::Right);
-    assert_eq!(b.get(23), ChildKind::Right);
-    assert_eq!(b.get(26), ChildKind::Right);
-    assert_eq!(b.get(27), ChildKind::Right);
-    assert_eq!(b.get(31), ChildKind::Left);
+    assert_eq!(b.bit(0), ChildKind::Right);
+    assert_eq!(b.bit(7), ChildKind::Left);
+    assert_eq!(b.bit(8), ChildKind::Left);
+    assert_eq!(b.bit(9), ChildKind::Right);
+    assert_eq!(b.bit(15), ChildKind::Left);
+    assert_eq!(b.bit(16), ChildKind::Right);
+    assert_eq!(b.bit(20), ChildKind::Right);
+    assert_eq!(b.bit(23), ChildKind::Right);
+    assert_eq!(b.bit(26), ChildKind::Right);
+    assert_eq!(b.bit(27), ChildKind::Right);
+    assert_eq!(b.bit(31), ChildKind::Left);
     let b2 = b.suffix(8);
     assert_eq!(b2.len(), 24);
-    assert_eq!(b2.get(0), ChildKind::Left);
-    assert_eq!(b2.get(1), ChildKind::Right);
-    assert_eq!(b2.get(7), ChildKind::Left);
-    assert_eq!(b2.get(12), ChildKind::Right);
-    assert_eq!(b2.get(15), ChildKind::Right);
+    assert_eq!(b2.bit(0), ChildKind::Left);
+    assert_eq!(b2.bit(1), ChildKind::Right);
+    assert_eq!(b2.bit(7), ChildKind::Left);
+    assert_eq!(b2.bit(12), ChildKind::Right);
+    assert_eq!(b2.bit(15), ChildKind::Right);
     let b3 = b2.suffix(24);
     assert_eq!(b3.len(), 0);
     let b4 = b.suffix(1);
-    assert_eq!(b4.get(6), ChildKind::Left);
-    assert_eq!(b4.get(7), ChildKind::Left);
-    assert_eq!(b4.get(8), ChildKind::Right);
+    assert_eq!(b4.bit(6), ChildKind::Left);
+    assert_eq!(b4.bit(7), ChildKind::Left);
+    assert_eq!(b4.bit(8), ChildKind::Right);
 }
 
 #[test]
-fn test_dbkey_truncate() {
-    let b = DBKey::read(b"\x00\x83wertyuiopasdfghjklzxcvbnm123456\x08");
+fn test_proof_path_prefix() {
+    let b = ProofPath::from_raw(*b"\x00\x83wertyuiopasdfghjklzxcvbnm123456\x08");
     assert_eq!(b.len(), 8);
-    assert_eq!(b.truncate(1).get(0), ChildKind::Right);
-    assert_eq!(b.truncate(1).len(), 1);
+    assert_eq!(b.prefix(1).bit(0), ChildKind::Right);
+    assert_eq!(b.prefix(1).len(), 1);
 }
 
 #[test]
-fn test_dbkey_len() {
-    let b = DBKey::read(b"\x01qwertyuiopasdfghjklzxcvbnm123456\x00");
+fn test_proof_path_len() {
+    let b = ProofPath::from_raw(*b"\x01qwertyuiopasdfghjklzxcvbnm123456\x00");
     assert_eq!(b.len(), 256);
 }
 
 #[test]
-#[should_panic(expected = "self.from + idx < self.to")]
-fn test_dbkey_at_overflow() {
-    let b = DBKey::read(b"\x00qwertyuiopasdfghjklzxcvbnm123456\x0F");
-    b.get(32);
+#[should_panic(expected = "self.start() + idx < self.end()")]
+fn test_proof_path_at_overflow() {
+    let b = ProofPath::from_raw(*b"\x00qwertyuiopasdfghjklzxcvbnm123456\x0F");
+    b.bit(32);
 }
 
 #[test]
-#[should_panic(expected = "self.from + suffix <= self.to")]
-fn test_dbkey_suffix_overflow() {
-    let b = DBKey::read(b"\x00qwertyuiopasdfghjklzxcvbnm123456\xFF");
+#[should_panic(expected = "pos <= self.end()")]
+fn test_proof_path_suffix_overflow() {
+    let b = ProofPath::from_raw(*b"\x00qwertyuiopasdfghjklzxcvbnm123456\xFF");
     assert_eq!(b"\x01qwertyuiopasdfghjklzxcvbnm123456\x00".len(), 34);
     b.suffix(255).suffix(2);
 }
 
 #[test]
-#[should_panic(expected = "self.from + idx < self.to")]
-fn test_dbkey_suffix_at_overflow() {
-    let b = DBKey::read(b"\x00qwertyuiopasdfghjklzxcvbnm123456\xFF");
-    b.suffix(1).get(255);
+#[should_panic(expected = "self.start() + idx < self.end()")]
+fn test_proof_path_suffix_bit_overflow() {
+    let b = ProofPath::from_raw(*b"\x00qwertyuiopasdfghjklzxcvbnm123456\xFF");
+    b.suffix(1).bit(255);
 }
 
 #[test]
-fn test_dbkey_common_prefix() {
-    let b1 = DBKey::read(b"\x01abcd0000000000000000000000000000\x00");
-    let b2 = DBKey::read(b"\x01abef0000000000000000000000000000\x00");
-    assert_eq!(b1.common_prefix(&b1), 256);
-    let c = b1.common_prefix(&b2);
+fn test_proof_path_common_prefix_len() {
+    let b1 = ProofPath::from_raw(*b"\x01abcd0000000000000000000000000000\x00");
+    let b2 = ProofPath::from_raw(*b"\x01abef0000000000000000000000000000\x00");
+    assert_eq!(b1.common_prefix_len(&b1), 256);
+    let c = b1.common_prefix_len(&b2);
     assert_eq!(c, 17);
-    let c = b2.common_prefix(&b1);
+    let c = b2.common_prefix_len(&b1);
     assert_eq!(c, 17);
     let b1 = b1.suffix(9);
     let b2 = b2.suffix(9);
-    let c = b1.common_prefix(&b2);
+    let c = b1.common_prefix_len(&b2);
     assert_eq!(c, 8);
-    let b3 = DBKey::read(b"\x01\xFF0000000000000000000000000000000\x00");
-    let b4 = DBKey::read(b"\x01\xF70000000000000000000000000000000\x00");
-    assert_eq!(b3.common_prefix(&b4), 3);
-    assert_eq!(b4.common_prefix(&b3), 3);
-    assert_eq!(b3.common_prefix(&b3), 256);
+    let b3 = ProofPath::from_raw(*b"\x01\xFF0000000000000000000000000000000\x00");
+    let b4 = ProofPath::from_raw(*b"\x01\xF70000000000000000000000000000000\x00");
+    assert_eq!(b3.common_prefix_len(&b4), 3);
+    assert_eq!(b4.common_prefix_len(&b3), 3);
+    assert_eq!(b3.common_prefix_len(&b3), 256);
     let b3 = b3.suffix(30);
-    assert_eq!(b3.common_prefix(&b3), 226);
-    let b3 = b3.truncate(200);
-    assert_eq!(b3.common_prefix(&b3), 200);
-    let b5 = DBKey::read(b"\x01\xF00000000000000000000000000000000\x00");
-    assert_eq!(b5.truncate(0).common_prefix(&b3), 0);
+    assert_eq!(b3.common_prefix_len(&b3), 226);
+    let b3 = b3.prefix(200);
+    assert_eq!(b3.common_prefix_len(&b3), 200);
+    let b5 = ProofPath::from_raw(*b"\x01\xF00000000000000000000000000000000\x00");
+    assert_eq!(b5.prefix(0).common_prefix_len(&b3), 0);
 }
 
 #[test]
-fn test_dbkey_is_leaf() {
-    let b = DBKey::read(b"\x01qwertyuiopasdfghjklzxcvbnm123456\x00");
+fn test_proof_path_is_leaf() {
+    let b = ProofPath::from_raw(*b"\x01qwertyuiopasdfghjklzxcvbnm123456\x00");
     assert_eq!(b.len(), 256);
     assert_eq!(b.suffix(4).is_leaf(), true);
     assert_eq!(b.suffix(8).is_leaf(), true);
     assert_eq!(b.suffix(250).is_leaf(), true);
-    assert_eq!(b.truncate(16).is_leaf(), false);
+    assert_eq!(b.prefix(16).is_leaf(), false);
 }
 
 #[test]
-fn test_dbkey_is_branch() {
-    let b = DBKey::read(b"\x00qwertyuiopasdfghjklzxcvbnm123456\xFF");
+fn test_proof_path_is_branch() {
+    let b = ProofPath::from_raw(*b"\x00qwertyuiopasdfghjklzxcvbnm123456\xFF");
     assert_eq!(b.len(), 255);
     assert_eq!(b.is_leaf(), false);
 }
 
 #[test]
-fn test_dbkey_debug_leaf() {
+fn test_proof_path_debug_leaf() {
     use std::fmt::Write;
-    let b = DBKey::read(b"\x01qwertyuiopasdfghjklzxcvbnm123456\x00");
+    let b = ProofPath::from_raw(*b"\x01qwertyuiopasdfghjklzxcvbnm123456\x00");
     let mut buf = String::new();
     write!(&mut buf, "{:?}", b).unwrap();
     assert_eq!(
         buf,
-        "DBKey { begin: 0, end: 256, bits: \"01110001|01110111|01100101|01110010|01110100|01111001\
-        |01110101|01101001|01101111|01110000|01100001|01110011|01100100|01100110|01100111|01101000\
-        |01101010|01101011|01101100|01111010|01111000|01100011|01110110|01100010|01101110|01101101\
-        |00110001|00110010|00110011|00110100|00110101|00110110|\" }"
+        "ProofPath { start: 0, end: 256, bits: \"01110001|01110111|01100101|01110010|01110100|0111\
+         1001|01110101|01101001|01101111|01110000|01100001|01110011|01100100|01100110|01100111|0110\
+         1000|01101010|01101011|01101100|01111010|01111000|01100011|01110110|01100010|01101110|0110\
+         1101|00110001|00110010|00110011|00110100|00110101|00110110|\" }"
     );
 }
 
 #[test]
-fn test_dbkey_debug_branch() {
+fn test_proof_path_debug_branch() {
     use std::fmt::Write;
-    let b = DBKey::read(b"\x00qwertyuiopasdfghjklzxcvbnm123456\xF0").suffix(12);
+    let b = ProofPath::from_raw(*b"\x00qwertyuiopasdfghjklzxcvbnm123456\xF0").suffix(12);
     let mut buf = String::new();
     write!(&mut buf, "{:?}", b).unwrap();
     assert_eq!(
         buf,
-        "DBKey { begin: 12, end: 240, bits: \"________|0111____|01100101|01110010|01110100|0111100\
-        1|01110101|01101001|01101111|01110000|01100001|01110011|01100100|01100110|01100111|0110100\
-        0|01101010|01101011|01101100|01111010|01111000|01100011|01110110|01100010|01101110|0110110\
-        1|00110001|00110010|00110011|00110100|________|________|\" }"
+        "ProofPath { start: 12, end: 240, bits: \"________|0111____|01100101|01110010|01110100|011\
+         11001|01110101|01101001|01101111|01110000|01100001|01110011|01100100|01100110|01100111|011\
+         01000|01101010|01101011|01101100|01111010|01111000|01100011|01110110|01100010|01101110|011\
+         01101|00110001|00110010|00110011|00110100|________|________|\" }"
     );
 }